ai_service/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
spacy
pytesseract
pillow
transformers
torch
